+++ OggConverter/src/Forms/QuickConvert.cs
@@ -1,4 +1,19 @@
-﻿using System;
+﻿// MSC Music Manager
+// Copyright(C) 2019 Athlon
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.If not, see<http://www.gnu.org/licenses/>.
+
 using System.Windows.Forms;
 using Microsoft.Win32;
 using System.IO;
@@ -41,16 +56,10 @@ namespace OggConverter
 
             this.files = files;
             Message = $"Where do you want to convert {files.Length} file{(files.Length > 1 ? "s" : "")}?";
-        }
 
-        private void BtnRadio_Click(object sender, EventArgs e)
-        {
-            Convert("Radio", 99);
-        }
-
-        private void BtnCD_Click(object sender, EventArgs e)
-        {
-            Convert("CD", 15);
+            btnRadio.Click += (s, e) => Convert("Radio", 99);
+            btnCD.Click += (s, e) => Convert("CD", 15);
+            btnExit.Click += (s, e) => Application.Exit();
         }
 
         async void Convert(string to, int limit)
@@ -67,11 +76,6 @@ namespace OggConverter
 
             Message = "Done!";
             btnExit.Visible = true;
-        }
-
-        private void BtnExit_Click(object sender, EventArgs e)
-        {
-            Application.Exit();
         }
     }
 }